#22 Main updates to BillOfMaterialItems model and traits, some updates to product traits
