+++ app/Models/BillOfMaterialItem.php
@@ -2,14 +2,16 @@
 
 namespace App\Models;
 
+use App\Concerns\BillOfMaterials\HasBOMItemAccessors;
+use App\Concerns\BillOfMaterials\HasBOMItemHelpers;
+use App\Concerns\BillOfMaterials\HasBOMItemScopes;
 use Database\Factories\BillOfMaterialItemFactory;
 use Illuminate\Database\Eloquent\Attributes\Fillable;
-use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Support\Carbon;
-use Illuminate\Database\Eloquent\Builder;
 
 #[Fillable([
     'bill_of_material_id',
@@ -45,7 +47,7 @@ use Illuminate\Database\Eloquent\Builder;
  * @property float|null $scrap_percentage
  * @property Carbon|null $created_at
  * @property Carbon|null $updated_at
- * 
+ *
  * @property-read BillOfMaterial $billOfMaterial
  * @property-read Product $product
  * @property-read Part $part
@@ -55,9 +57,15 @@ class BillOfMaterialItem extends Model
     /**
      * @use HasFactory<BillOfMaterialItemFactory>
      * @use SoftDeletes<SoftDeletes>
+     * @use HasBOMItemAccessors<HasBOMItemAccessors>
+     * @use HasBOMItemHelpers<HasBOMItemHelpers>
+     * @use HasBOMItemScopes<HasBOMItemScopes>
      */
     use HasFactory,
-        SoftDeletes;
+        SoftDeletes,
+        HasBOMItemAccessors,
+        HasBOMItemHelpers,
+        HasBOMItemScopes;
 
     /**
      * Get the user who created the pipeline.
@@ -121,195 +129,6 @@ class BillOfMaterialItem extends Model
     public function part(): BelongsTo
     {
         return $this->belongsTo(Part::class);
-    }
-
-    // ==================== SCOPES ====================
-    /**
-     * Scope a query to only include real BOM Item.
-     *
-     * @param  Builder $query
-     *
-     * @return Builder
-     */
-    public function scopeReal(Builder $query): Builder
-    {
-        return $query->where('is_real', true);
-    }
-
-    /**
-     * Scope: Required items only (not optional).
-     *
-     * @param  Builder $query
-     *
-     * @return Builder
-     */
-    public function scopeRequired(Builder $query): Builder
-    {
-        return $query->where('is_optional', false);
-    }
-
-    /**
-     * Scope: Optional items only.
-     *
-     * @param  Builder $query
-     *
-     * @return Builder
-     */
-    public function scopeOptional(Builder $query): Builder
-    {
-        return $query->where('is_optional', true);
-    }
-
-    /**
-     * Scope: Order by sequence.
-     *
-     * @param  Builder $query
-     *
-     * @return Builder
-     */
-    public function scopeOrdered(Builder $query): Builder
-    {
-        return $query->orderBy('sequence');
-    }
-
-    // ==================== ACCESSORS ====================
-
-    /**
-     * Calculate the total cost for this BOM item.
-     *
-     * @return float
-     */
-    public function getTotalCostAttribute(): float
-    {
-        $partCost = $this->part->cost_price ?? 0;
-        $quantityWithScrap = $this->getQuantityWithScrap();
-        
-        return $quantityWithScrap * $partCost;
-    }
-
-    /**
-     * Get formatted total cost.
-     *
-     * @return string
-     */
-    public function getFormattedTotalCostAttribute(): string
-    {
-        $currency = $this->product->currency ?? 'GBP';
-        $symbols = [
-            'GBP' => '£',
-            'USD' => '$',
-            'EUR' => '€',
-        ];
-        
-        $symbol = $symbols[$currency] ?? $currency . ' ';
-        
-        return $symbol . number_format($this->total_cost, 2);
-    }
-
-    /**
-     * Get quantity including scrap allowance.
-     *
-     * @return float
-     */
-    public function getQuantityWithScrapAttribute(): float
-    {
-        return $this->getQuantityWithScrap();
-    }
-
-    /**
-     * Check if part is available in sufficient quantity.
-     *
-     * @return bool
-     */
-    public function getIsAvailableAttribute(): bool
-    {
-        return $this->part->quantity >= $this->quantity;
-    }
-
-    /**
-     * Get shortage quantity if part is not available.
-     *
-     * @return float
-     */
-    public function getShortageAttribute(): float
-    {
-        $shortage = $this->quantity - $this->part->quantity;
-        return max(0, $shortage);
-    }
-
-    // ==================== HELPER METHODS ====================
-
-    /**
-     * Calculate quantity including scrap percentage.
-     *
-     * @return float
-     */
-    public function getQuantityWithScrap(): float
-    {
-        if ($this->scrap_percentage === null || $this->scrap_percentage == 0) {
-            return $this->quantity;
-        }
-
-        return $this->quantity * (1 + ($this->scrap_percentage / 100));
-    }
-
-    /**
-     * Calculate quantity needed for multiple products.
-     *
-     * @param  int $productQuantity
-     *
-     * @return float
-     */
-    public function calculateQuantityForProducts(int $productQuantity): float
-    {
-        return $this->getQuantityWithScrap() * $productQuantity;
-    }
-
-    /**
-     * Check if sufficient stock exists for manufacturing.
-     *
-     * @param  int $productQuantity
-     *
-     * @return array
-     */
-    public function checkStockAvailability(int $productQuantity = 1): array
-    {
-        $requiredQuantity = $this->calculateQuantityForProducts($productQuantity);
-        $availableQuantity = $this->part->quantity;
-        $sufficient = $availableQuantity >= $requiredQuantity;
-
-        return [
-            'part_id' => $this->part_id,
-            'part_name' => $this->part->name,
-            'part_sku' => $this->part->sku,
-            'required_per_product' => $this->quantity,
-            'scrap_percentage' => $this->scrap_percentage,
-            'quantity_with_scrap' => $this->getQuantityWithScrap(),
-            'total_required' => $requiredQuantity,
-            'available' => $availableQuantity,
-            'sufficient' => $sufficient,
-            'shortage' => max(0, $requiredQuantity - $availableQuantity),
-        ];
-    }
-
-    /**
-     * Get extended cost (quantity × unit cost).
-     *
-     * @return float
-     */
-    public function getExtendedCost(): float
-    {
-        return $this->total_cost;
-    }
-
-    /**
-     * Get formatted extended cost.
-     *
-     * @return string
-     */
-    public function getFormattedExtendedCost(): string
-    {
-        return $this->formatted_total_cost;
     }
 
     /**
